% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{calcule_erreurs}
\alias{calcule_erreurs}
\title{calcule_erreurs}
\usage{
calcule_erreurs(r, phat)
}
\arguments{
\item{r}{vector containing values of the response dummy variable}

\item{phat}{vector containing values of estimated response probability}
}
\description{
calcule_erreurs
}
