% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{calcule_variances}
\alias{calcule_variances}
\title{calcule_variances}
\usage{
calcule_variances(phat, x, p)
}
\arguments{
\item{phat}{vector containing values of estimated response probability}

\item{x}{vector containing values of the variable of interest}

\item{p}{value of the global sample share used as a test sample}
}
\description{
calcule_variances
}
