% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidation.R
\name{choose_best_hb}
\alias{choose_best_hb}
\title{choose_best_hb}
\usage{
choose_best_hb(y, phat, weight = rep(1, length(y)), u, parameters,
  method = "mse", K = 10, print_rhg = F)
}
\arguments{
\item{y}{vector}

\item{phat}{vector containing estimated response probabilities}

\item{weight}{vector containing the weights associated to each observation}

\item{u}{vector containing}

\item{parameters}{v}

\item{method}{v}

\item{K}{pouet}

\item{print_rhg}{pouetpouet}
}
\description{
choose_best_hb
}
