% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb.R
\name{hb}
\alias{hb}
\title{blabla}
\usage{
hb(phat, parameters, leaf = 100, print_rhg = T)
}
\arguments{
\item{phat}{bloblo}

\item{parameters}{bloblo}

\item{leaf}{bloblo}

\item{print_rhg}{bloblo}
}
\value{
bloblo
}
\description{
blibli
}
